backport refactor for splitting perses and legacy dashboards
