+++ web/src/components/dashboards/legacy/table.tsx
@@ -16,7 +16,7 @@ import * as _ from 'lodash-es';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import ErrorAlert from './error';
+import ErrorAlert from '../shared/error';
 import { getPrometheusURL } from '../../console/graphs/helpers';
 import { usePoll } from '../../console/utils/poll-hook';
 import { useSafeFetch } from '../../console/utils/safe-fetch-hook';
