+++ web/src/components/dashboards/shared/error.tsx
